+++ CHANGES.md
@@ -4,6 +4,7 @@ New languages:
 
 - *pf.conf* by [Peter Piwowarski][]
 - *Julia* by [Kenta Sato][]
+- *Inform7* by [Bruno Dias][]
 
 New Styles: